add book_router and model Book
